+++ apps/docs/src/app/core/component-docs/datetime-picker/datetime-picker-docs.module.ts
@@ -13,7 +13,7 @@ import { DatetimeFormatExampleComponent } from './examples/datetime-format-examp
 import { DatetimeFormExampleComponent } from './examples/datetime-form-example/datetime-form-example.component';
 import { DatetimePickerAllowNullExampleComponent } from './examples/datetime-allow-null-example/datetime-allow-null-example.component';
 import { DatetimeDisabledExampleComponent } from './examples/datetime-disabled-example/datetime-disabled-example.component';
-import { SegmentedButtonModule, DatetimePickerModule } from '@fundamental-ngx/core';
+import { DatetimePickerModule, SelectModule } from '@fundamental-ngx/core';
 
 const routes: Routes = [
     {
@@ -27,7 +27,7 @@ const routes: Routes = [
 ];
 
 @NgModule({
-    imports: [RouterModule.forChild(routes), SharedDocumentationModule, DatetimePickerModule, SegmentedButtonModule],
+    imports: [RouterModule.forChild(routes), SharedDocumentationModule, DatetimePickerModule, SelectModule],
     exports: [RouterModule],
     declarations: [
         DatetimeExampleComponent,