+++ apps/docs/src/app/core/component-docs/time-picker/time-picker-docs.module.ts
@@ -8,13 +8,14 @@ import { TimePickerDocsComponent } from './time-picker-docs.component';
 import { TimePickerExampleComponent } from './examples/time-picker-example.component';
 import { TimePicker12ExampleComponent } from './examples/time-picker-12-example.component';
 import { TimePickerFormExampleComponent } from './examples/time-picker-form-example.component';
+import { TimePickerLocaleExampleComponent } from './examples/time-picker-locale-example/time-picker-locale-example.component';
 import { TimePickerCompactExampleComponent } from './examples/time-picker-compact-example.component';
 import { TimePickerDisabledExampleComponent } from './examples/time-picker-disabled-example.component';
 import { TimePickerOnlyHoursExampleComponent } from './examples/time-picker-only-hours-example.component';
 import { TimePickerNoSecondsExampleComponent } from './examples/time-picker-no-seconds-example.component';
 import { TimePickerAllowNullExampleComponent } from './examples/time-picker-allow-null-example.component';
 import { TimePickerOtherDelimiterExampleComponent } from './examples/time-picker-other-delimiter-example.component';
-import { TimePickerModule, SegmentedButtonModule } from '@fundamental-ngx/core';
+import { TimePickerModule, SegmentedButtonModule, SelectModule } from '@fundamental-ngx/core';
 
 const routes: Routes = [
     {
@@ -28,7 +29,7 @@ const routes: Routes = [
 ];
 
 @NgModule({
-    imports: [RouterModule.forChild(routes), SharedDocumentationModule, TimePickerModule, SegmentedButtonModule],
+    imports: [RouterModule.forChild(routes), SharedDocumentationModule, TimePickerModule, SegmentedButtonModule, SelectModule],
     exports: [RouterModule],
     declarations: [
         TimePickerDocsComponent,
@@ -36,6 +37,7 @@ const routes: Routes = [
         TimePickerExampleComponent,
         TimePicker12ExampleComponent,
         TimePickerFormExampleComponent,
+        TimePickerLocaleExampleComponent,
         TimePickerCompactExampleComponent,
         TimePickerDisabledExampleComponent,
         TimePickerOnlyHoursExampleComponent,