+++ CHANGELOG.md
@@ -2,6 +2,48 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.21.1-rc.2](https://github.com/SAP/fundamental-ngx/compare/v0.21.1-rc.1...v0.21.1-rc.2) (2020-08-14)
+
+### [0.21.1-rc.1](https://github.com/SAP/fundamental-ngx/compare/v0.21.1-rc.0...v0.21.1-rc.1) (2020-08-14)
+
+### [0.21.1-rc.0](https://github.com/SAP/fundamental-ngx/compare/v0.21.0...v0.21.1-rc.0) (2020-08-14)
+
+
+### Features
+
+* refactor DND directive ([#2996](https://github.com/SAP/fundamental-ngx/issues/2996)) ([9e6d021](https://github.com/SAP/fundamental-ngx/commit/9e6d0217622eef2cf0994f1d6c08066bc9e91275))
+
+## [0.21.0](https://github.com/SAP/fundamental-ngx/compare/v0.21.0-rc.10...v0.21.0) (2020-08-13)
+
+## [0.21.0-rc.10](https://github.com/SAP/fundamental-ngx/compare/v0.21.0-rc.9...v0.21.0-rc.10) (2020-08-13)
+
+
+### Bug Fixes
+
+* add locale to i18n example of datetime picker ([#2923](https://github.com/SAP/fundamental-ngx/issues/2923)) ([3920897](https://github.com/SAP/fundamental-ngx/commit/3920897e1be706ef7821cbd66191307258c88b5d))
+
+## [0.21.0-rc.9](https://github.com/SAP/fundamental-ngx/compare/v0.21.0-rc.8...v0.21.0-rc.9) (2020-08-13)
+
+
+### Bug Fixes
+
+* remove time disabled example, fix combobox in shellbar ([#3025](https://github.com/SAP/fundamental-ngx/issues/3025)) ([22f0fab](https://github.com/SAP/fundamental-ngx/commit/22f0fab921a44fd0ee46ad6c203a23cf01bf99b3))
+
+## [0.21.0-rc.8](https://github.com/SAP/fundamental-ngx/compare/v0.21.0-rc.7...v0.21.0-rc.8) (2020-08-12)
+
+## [0.21.0-rc.7](https://github.com/SAP/fundamental-ngx/compare/v0.21.0-rc.6...v0.21.0-rc.7) (2020-08-11)
+
+## [0.21.0-rc.6](https://github.com/SAP/fundamental-ngx/compare/v0.21.0-rc.5...v0.21.0-rc.6) (2020-08-10)
+
+## [0.21.0-rc.5](https://github.com/SAP/fundamental-ngx/compare/v0.21.0-rc.4...v0.21.0-rc.5) (2020-08-10)
+
+## [0.21.0-rc.4](https://github.com/SAP/fundamental-ngx/compare/v0.21.0-rc.3...v0.21.0-rc.4) (2020-08-10)
+
+
+### Bug Fixes
+
+* add time picker with locale example ([#2925](https://github.com/SAP/fundamental-ngx/issues/2925)) ([359f81c](https://github.com/SAP/fundamental-ngx/commit/359f81cb8bef8b52df6b0db603f8de80be0b7c1c))
+
 ## [0.21.0-rc.3](https://github.com/SAP/fundamental-ngx/compare/v0.21.0-rc.2...v0.21.0-rc.3) (2020-08-07)
 
 ## [0.21.0-rc.2](https://github.com/SAP/fundamental-ngx/compare/v0.21.0-rc.1...v0.21.0-rc.2) (2020-08-07)